requirements.txt
numpy==1.26.0
opencv-python==4.7.0.68
PyQt5==5.15.7
scikit-learn>=1.0.2
matplotlib==3.5.3
PyYAML>=6.0
pandas>=1.5.3
av==10.0.0
more_itertools==8.12.0